pyproject.toml
[tool.poetry]
name = "flask-book-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "flask_book_app"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.3.2"
flask-sqlalchemy = "^3.0.3"
python-dotenv = "^1.0.0"
pymysql = "^1.0.3"
flask-wtf = "^1.1.1"
flask-login = "^0.6.2"
flask-bcrypt = "^1.0.1"
email-validator = "^2.0.0.post2"
psycopg2-binary = "^2.9.6"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
